net, grpc_support: Set NetworkIsolationKey from header
If BidirectionalStream request contains a -network-isolation-key
header, it is used to set the network isolation key of the stream.
The header itself is removed and not transmitted.

The header value should be a valid URL with different host and port
for each different network isolation key. Invalid header value is
reported by returning error from bidirectional_stream_start.

Network isolation takes effect only if it is enabled by experimental
option of

  "feature_list": {
    "enable-features": "PartitionConnectionsByNetworkIsolationKey"
  }